requirements.txt
flask==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-socketio==5.5.1
netmiko==4.5.0
paramiko==3.5.1
ping3==4.0.8
werkzeug==3.1.3
gevent==24.11.1
gevent-websocket==0.10.1
cryptography==42.0.5
python-dotenv==1.0.1
APScheduler==3.10.4
